first

+++ include/laextension.h
@@ -1,8 +1,8 @@
 #ifndef LAEXTENSION
 #define LAEXTENSION
-void cross(float* A, float* B, float* C);
-float min_arr(float* A, int elements);
-float max_arr(float* A, int elements);
-float trace(float* A, int side);
-float l2_norm(float* A, int size);
+void cross(double* A, double* B, double* C);
+double min_arr(double* A, int elements);
+double max_arr(double* A, int elements);
+double trace(double* A, int side);
+double l2_norm(double* A, int size);
 #endif
