+++ linalg/LinearAlgebra/diag.c
@@ -10,10 +10,10 @@
  * Create a diagonal matrix B with the size row x column, of a vector A.
  * Notice that the row of vector A need to be the same length as the column of matrix B.
  */
-void diag(float* A, float* B, int row_b, int column_b) {
+void diag(double* A, double* B, int row_b, int column_b) {
 
     // Reset B
-    memset(B, 0, row_b * column_b * sizeof(float));
+    memset(B, 0, row_b * column_b * sizeof(double));
 
     for (int i = 0; i < row_b; i++) {
         for (int j = 0; j < column_b; j++) {
