+++ linalg/LinearAlgebra/eabs.c
@@ -9,7 +9,7 @@
 /*
  * Take the fabsolue values of every element of matrix A, size row x column
  */
-void efabs(float* A, int row, int column) {
+void efabs(double* A, int row, int column) {
     for (int i = 0; i < row; i++)
         for (int j = 0; j < column; j++) {
             *A = fabs(*(A));
