+++ linalg/LinearAlgebra/inv.c
@@ -11,20 +11,20 @@
  * Turn A into A^-1. A have the size row x row
  */
 
-void inv(float* A, int row) {
+void inv(double* A, int row) {
 
     // Create identity matrix
-    float I[row * row];
+    double I[row * row];
     eye(I, row, row);
 
     // Do inverse of A
     linsolve(A, A, I, row, row);
 }
 
-void invf(float* A, int row) {
+void invf(double* A, int row) {
 
     // Create identity matrix
-    float I[row * row];
+    double I[row * row];
     eye(I, row, row);
 
     // Do inverse of A
