+++ linalg/LinearAlgebra/move.c
@@ -12,10 +12,10 @@
  * Matrix A have the size row x column
  */
 
-void move(float* A, int row, int column, int down, int right) {
+void move(double* A, int row, int column, int down, int right) {
 
-    float B[row * column];
-    memset(B, 0, row * column * sizeof(float));
+    double B[row * column];
+    memset(B, 0, row * column * sizeof(double));
 
     /*
      * We take -1 due to indexing from zero
@@ -30,5 +30,5 @@ void move(float* A, int row, int column, int down, int right) {
         }
     }
 
-    memcpy(A, B, row * column * sizeof(float));
+    memcpy(A, B, row * column * sizeof(double));
 }
