+++ linalg/LinearAlgebra/print.c
@@ -9,7 +9,7 @@
 /*
  * Print a matrix A, with the dimension row x column
  */
-void print(float* A, int row, int column) {
+void print(double* A, int row, int column) {
     for (int i = 0; i < row; i++) {
         for (int j = 0; j < column; j++) {
             printf("%0.18f ", *(A++));
@@ -28,7 +28,7 @@ void print(float* A, int row, int column) {
  * \param row the number of rows in the matrix
  * \param column the number of columns in the matrix
  */
-void debug_matrix(float* A, int row, int column) {
+void debug_matrix(double* A, int row, int column) {
     for (int i = 0; i < row; i++) {
         for (int j = 0; j < column; j++) {
             printf("%0.18f ", *(A++));
