+++ linalg/LinearAlgebra/sub.c
@@ -12,7 +12,7 @@
  * row_a x column_a
  */
 
-void sub(float* A, float* B, float* C, int row_a, int column_a, int column_b) {
+void sub(double* A, double* B, double* C, int row_a, int column_a, int column_b) {
 
     /*
      * This uses the formula C=A-B
