+++ linalg/LinearAlgebra/tran.c
@@ -11,11 +11,11 @@
  * Take the transpose of a matrix A, with the dimension row x column. The result will be A' with the
  * dimension column x row
  */
-void tran(float* A, int row, int column) {
+void tran(double* A, int row, int column) {
 
-    float B[row * column];
-    float* transpose;
-    float* ptr_A = A;
+    double B[row * column];
+    double* transpose;
+    double* ptr_A = A;
 
     for (int i = 0; i < row; i++) {
         transpose = &B[i];
@@ -27,14 +27,14 @@ void tran(float* A, int row, int column) {
     }
 
     // Copy!
-    memcpy(A, B, row * column * sizeof(float));
+    memcpy(A, B, row * column * sizeof(double));
 }
 
-void tranf(float* A, int row, int column) {
+void tranf(double* A, int row, int column) {
 
-    float B[row * column];
-    float* transpose;
-    float* ptr_A = A;
+    double B[row * column];
+    double* transpose;
+    double* ptr_A = A;
 
     for (int i = 0; i < row; i++) {
         transpose = &B[i];
@@ -46,5 +46,5 @@ void tranf(float* A, int row, int column) {
     }
 
     // Copy!
-    memcpy(A, B, row * column * sizeof(float));
+    memcpy(A, B, row * column * sizeof(double));
 }
