+++ linalg/LinearAlgebra/zeros.c
@@ -9,4 +9,4 @@
  * Turn all elements of A, size row x column, into 0.
  */
 
-void zeros(float* A, int row, int column) { memset(A, 0, row * column * sizeof(float)); }
+void zeros(double* A, int row, int column) { memset(A, 0, row * column * sizeof(double)); }
