+++ src/laextension.c
@@ -1,18 +1,18 @@
 #include "include/laextension.h"
 #include "math.h"
 /*Cross Product for 2 1x3 vectors*/
-void cross(float* A, float* B, float* C){
+void cross(double* A, double* B, double* C){
     C[0] = A[1]*B[2] - A[2]*B[1];
     C[1] = A[2]*B[0] - A[0]*B[2];
     C[2] = A[0]*B[1] - A[1]*B[0];
 }
 /*Max*/
-float max_arr(float* A, int elements){
+double max_arr(double* A, int elements){
     if(elements == 0){
         return 0;
     }
     else{
-        float max_found = A[0];
+        double max_found = A[0];
         for(int i = 0; i<elements; i++){
             max_found = fmaxf(max_found, A[i]);
         }
@@ -20,12 +20,12 @@ float max_arr(float* A, int elements){
     }
 }
 /*Mix*/
-float min_arr(float* A, int elements){
+double min_arr(double* A, int elements){
     if(elements == 0){
         return 0;
     }
     else{
-        float min_found = A[0];
+        double min_found = A[0];
         for(int i = 0; i<elements; i++){
             min_found = fminf(min_found, A[i]);
         }
@@ -33,18 +33,18 @@ float min_arr(float* A, int elements){
     }
 }
 /*Trace*/
-float trace(float* A, int sz){
-    float n= 0;
+double trace(double* A, int sz){
+    double n= 0;
     for(int i = 0; i<sz; i++){
         n += A[i * sz + i];
     }
     return n;
 }
 /*L2Norm*/
-float l2_norm(float* A, int n){
-    float x = 0;
+double l2_norm(double* A, int n){
+    double x = 0;
     for (int i = 0; i<n; i++){
         x += pow(A[i], 2);
     }
-    return sqrtf(x);
+    return sqrt(x);
 }
